Refactoring: Move calc functions to container components.
